Kill some warnings, setting up new workspace

+++ src/main/java/com/bewitchment/client/render/entity/model/ModelWerewolf.java
@@ -466,10 +466,10 @@ public class ModelWerewolf extends ModelBase {
 						bipedLeftLeg.rotateAngleX = -1.2217304764F + 0.52359877559F * MathHelper.cos(time_running);
 						bipedRightLeg.rotateAngleX = -1.2217304764F + 0.52359877559F * MathHelper.cos((float) (time_running + 0.5));
 						lLeg02.rotateAngleX = 1.0471975512F + 0.52359877559F * MathHelper.sin((float) (time_running + 4.5));
-						rLeg02.rotateAngleX = 1.0471975512F + 0.52359877559F * MathHelper.sin((float) (time_running + 5));
+						rLeg02.rotateAngleX = 1.0471975512F + 0.52359877559F * MathHelper.sin(time_running + 5);
 						lFoot.rotateAngleX = 0.3839724354387525f + 0.43633231299F * MathHelper.cos(time_running);
 						rFoot.rotateAngleX = 0.3839724354387525f + 0.43633231299F * MathHelper.cos((float) (time_running + 0.5));
-						bipedBody.setRotationPoint(0.0F, (float) (-1.5F + 5 * MathHelper.sin((float) (time_running + 4.5))), 6.0F);
+						bipedBody.setRotationPoint(0.0F, -1.5F + 5 * MathHelper.sin((float) (time_running + 4.5)), 6.0F);
 						tail01Floof.rotateAngleX = 0.34906585039F + 0.26179938779F * MathHelper.cos(time_running);
 						tail02Floof.rotateAngleX = -0.3490658503988659F + 0.26179938779F * MathHelper.sin(time_running);
 						tail03Floof.rotateAngleX = -0.17453292519943295F + 0.08726646259F * MathHelper.sin(time_running);
